% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{convertdss}
\alias{convertdss}
\alias{convertdss-package}
\title{convertdss: A package for converting dss into other data formats}
\description{
The convertdss package provides functionality to convert dss files into:
NetCDF, SQLite.
}
\section{NetCDF}{

About NetCDF conversion
}

\section{SQLite}{

About SQLite conversion
}

