% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dss_to_sqlite_init}
\alias{dss_to_sqlite_init}
\title{Initilizes a sqlite database to recieve dss data}
\usage{
dss_to_sqlite_init(dss, db_file = "convertdss.db", overwrite = TRUE)
}
\arguments{
\item{dss}{a dss file object from \code{\link[dssrip]{opendss}}}

\item{db_file}{}

\item{overwrite}{}
}
\value{
An SQLite file handle
}
\description{
Uses an open dss file. A time dimension is defined with a corresponding
}
\details{
It is important that all the times be defined before any data gets written,
}
\author{
Cameron Bracken
}

