% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dss_var_to_sqlite}
\alias{dss_var_to_sqlite}
\title{Writes a dss variable to an existing sqlite file}
\usage{
dss_var_to_sqlite(v, db)
}
\arguments{
\item{v}{a variable object returned by \code{\link{read_dss_variable}}}

\item{db}{A RSQLite file handle set up by \code{\link{dss_to_sqlite_init}}}
}
\value{
TRUE if the write was successful, false otherwise
}
\description{
Uses the output of read_dss_variable, creating a new table in an _existing_
sqlite file
}
\author{
Cameron Bracken
}

